‘代码同步’

+++ trace_streamer/src/parser/rawtrace_parser/cpu_detail_parser.cpp
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) 2021 Huawei Device Co., Ltd.
+ * Copyright (c) Huawei Technologies Co., Ltd. 2023. All rights reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -263,9 +263,9 @@ bool CpuDetailParser::SchedSwitchEvent(const RawTraceEventInfo& event)
     auto uprevtid = streamFilters_->processFilter_->UpdateOrCreateThreadWithName(
         event.msgPtr->timestamp(), schedSwitchMsg.prev_pid(), schedSwitchMsg.prev_comm());
 
-    streamFilters_->cpuFilter_->InsertSwitchEvent(
-        event.msgPtr->timestamp(), event.cpuId, uprevtid, static_cast<uint64_t>(schedSwitchMsg.prev_prio()),
-        schedSwitchPrevState, nextInternalTid, static_cast<uint64_t>(schedSwitchMsg.next_prio()), INVALID_DATAINDEX);
+    streamFilters_->cpuFilter_->InsertSwitchEvent(event.msgPtr->timestamp(), event.cpuId, uprevtid,
+                                                  schedSwitchMsg.prev_prio(), schedSwitchPrevState, nextInternalTid,
+                                                  schedSwitchMsg.next_prio(), INVALID_DATAINDEX);
     return true;
 }
 bool CpuDetailParser::SchedBlockReasonEvent(const RawTraceEventInfo& event)
