+++ trace_streamer/src/trace_data/trace_stdtype/ftrace/sched_stdtype.h
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) 2021 Huawei Device Co., Ltd.
+ * Copyright (c) Huawei Technologies Co., Ltd. 2023. All rights reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *     http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -92,7 +92,7 @@ public:
                             uint64_t cpu,
                             uint32_t internalTid,
                             uint64_t endState,
-                            uint64_t priority);
+                            int32_t priority);
     void SetDuration(size_t index, uint64_t duration);
     void Update(uint64_t index, uint64_t ts, uint64_t state);
     void UpdateEndState(uint64_t index, uint64_t state);
@@ -103,7 +103,7 @@ public:
         return endStates_;
     }
 
-    const std::deque<uint64_t>& PriorityData() const
+    const std::deque<int32_t>& PriorityData() const
     {
         return priority_;
     }
@@ -145,7 +145,7 @@ private:
     std::deque<InternalPid> internalPids_ = {};
     std::deque<uint64_t> tsEnds_ = {};
     std::deque<uint64_t> endStates_ = {};
-    std::deque<uint64_t> priority_ = {};
+    std::deque<int32_t> priority_ = {};
     std::deque<uint32_t> argSets_ = {};
 };
 class Raw : public CacheBase, public BatchCacheBase {
